+++ Network/API/Telegram/Bot/Object/Update/Message/Content/File/Video.hs
@@ -1,7 +1,6 @@
 module Network.API.Telegram.Bot.Object.Update.Message.Content.File.Video (Video (..)) where
 
-import "aeson" Data.Aeson (FromJSON (parseJSON)
-	, object, withObject, (.:), (.:?), (.=))
+import "aeson" Data.Aeson (FromJSON (parseJSON), withObject, (.:), (.:?))
 import "base" Control.Applicative ((<*>))
 import "base" Data.Int (Int)
 import "base" Data.Maybe (Maybe)
@@ -10,13 +9,9 @@ import "base" Data.Functor ((<$>))
 import "base" Text.Show (Show)
 import "text" Data.Text (Text)
 
-import Network.API.Telegram.Bot.Object.Update.Message.Content.File.Size (Size)
-
-data Video = Video Int Int Int (Maybe Size) (Maybe Text) (Maybe Int)
+data Video = Video Int Int Int (Maybe Text) (Maybe Int)
 	deriving Show
 
 instance FromJSON Video where
-	parseJSON = withObject "Video" $ \v -> Video
-		<$> v .: "width" <*> v .: "height"
-		<*> v .: "duration" <*> v .:?  "thumb"
-		<*> v .:? "mime_type" <*> v .:? "file_size"
+	parseJSON = withObject "Video" $ \v -> Video <$> v .: "width" <*> v .: "height"
+		<*> v .: "duration" <*> v .:? "mime_type" <*> v .:? "file_size"